+++ tests/PSRule.Rules.Azure.Tests/Azure.FrontDoor.Tests.ps1
@@ -274,7 +274,7 @@ Describe 'Azure.FrontDoor' -Tag 'Network', 'FrontDoor' {
     Context 'With template' {
         $templatePath = Join-Path -Path $here -ChildPath 'Resources.Template3.json';
         $outputFile = Join-Path -Path $rootPath -ChildPath out/tests/Resources.FrontDoor.json;
-        Export-AzTemplateRuleData -TemplateFile $templatePath -OutputPath $outputFile;
+        Export-AzRuleTemplateData -TemplateFile $templatePath -OutputPath $outputFile;
         $result = Invoke-PSRule -Module PSRule.Rules.Azure -InputPath $outputFile -Outcome All -WarningAction Ignore -ErrorAction Stop;
 
         It 'Azure.FrontDoor.State' {