+++ tests/PSRule.Rules.Azure.Tests/Azure.Policy.Tests.ps1
@@ -56,7 +56,7 @@ Describe 'Azure.Policy' {
         $templatePath = Join-Path -Path $here -ChildPath 'Resources.Policy.Template.json';
         $parameterPath = Join-Path -Path $here -ChildPath 'Resources.Policy.Parameters.json';
         $outputFile = Join-Path -Path $rootPath -ChildPath out/tests/Resources.Policy.json;
-        Export-AzTemplateRuleData -TemplateFile $templatePath -ParameterFile $parameterPath -OutputPath $outputFile;
+        Export-AzRuleTemplateData -TemplateFile $templatePath -ParameterFile $parameterPath -OutputPath $outputFile;
         $result = Invoke-PSRule -Module PSRule.Rules.Azure -InputPath $outputFile -Outcome All -WarningAction Ignore -ErrorAction Stop;
 
         It 'Azure.Policy.Descriptors' {